+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased](https://github.com/rerun-io/rerun/compare/latest...HEAD)
 * Add support for invisible tiles
+* `PartialEq` for `Tiles` now ignore internal state
 
 
 ## 0.1.0 - Initial Release - 2023-05-24